MyCineMusic/requirements.txt
streamlit>=1.33
psycopg2-binary>=2.9
pandas>=2.0
python-dotenv>=1.0
numpy<2
